Add components for enhanced component management
Introduce several new components for the editor's component tab, including `ComponentSearch`, `ComponentsList`, `ComponentItem`, and `LoadingComponents`. These components facilitate searching, displaying, and managing components within the editor. Additionally, a new `SearchProvider` is implemented to manage search state effectively, replacing the previous search context. This update enhances user experience by providing a structured and interactive way to handle components.
