+++ packages/packages/libs/utils/src/use-search-state.ts
@@ -1,0 +1,24 @@
+import { useDebounceState } from '../use-debounce-state';
+
+export function useSearchState( { localStorageKey }: { localStorageKey: string } ) {
+	const getInitialSearchValue = () => {
+		if ( localStorageKey ) {
+			const storedValue = localStorage.getItem( localStorageKey );
+			if ( storedValue ) {
+				localStorage.removeItem( localStorageKey );
+				return storedValue;
+			}
+		}
+		return '';
+	};
+	const { debouncedValue, inputValue, handleChange, setInputValue } = useDebounceState( {
+		delay: 300,
+		initialValue: getInitialSearchValue(),
+	} );
+	return {
+		debouncedValue,
+		inputValue,
+		handleChange,
+		setInputValue,
+	};
+}
